requirement.txt
pyautogen
pyautogen[websurfer]
pyautogen[retrievechat]
pyautogen[teachable]